Tariffs

+++ webapp/shared/datex2/v3_5_json_realtime/models/duration_value_input.py
@@ -16,7 +16,7 @@ class DurationValueInput(ValidataclassMixin):
     A measured or calculated value of a period of time.
     """
 
-    duration: int = FloatValidator()
+    duration: float = FloatValidator()
     comDataValueExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
