+++ webapp/shared/datex2/v3_5_json_static/models/heaviest_axle_weight_characteristic_input.py
@@ -18,7 +18,7 @@ class HeaviestAxleWeightCharacteristicInput(ValidataclassMixin):
     """
 
     comparisonOperator: ComparisonOperatorEnumGInput = DataclassValidator(ComparisonOperatorEnumGInput)
-    heaviestAxleWeight: int = FloatValidator()
+    heaviestAxleWeight: float = FloatValidator()
     comHeaviestAxleWeightCharacteristicExtensionG: ExtensionTypeGInput | UnsetValueType = (
         DataclassValidator(ExtensionTypeGInput),
         Default(UnsetValue),
